app/llm-gemma-variant/pyproject.toml
[tool.poetry]
name = "llm-gemma-variant"
version = "0.1.0"
description = ""
authors = ["Tacoman99 <jorgepadres789@gmail.com>"]
readme = "README.md"
packages = [{include = "llm_gemma_variant", from = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
llama-index-vector-stores-weaviate = "^1.1.3"
llama-index = "^0.11.21"
weaviate-client = "^4.9.3"
python-dotenv = "^1.0.1"
llama-index-llms-nvidia = "^0.2.6"
loguru = "^0.7.2"
torch = "^2.5.1"
llama-index-llms-ollama = "^0.3.4"
llama-index-llms-huggingface-api = "^0.2.0"
llama-index-llms-huggingface = "^0.3.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
